// src/app/shared/models/IProduct.ts
export interface IProduct {
  id: number;
  category: string;
  lensColor: string;
  lensWidth: number;
  lensHeight: number;
  lensMaterial: string;
  totalWidth: number;
  bracketLength: number;
  frameColor: string;
  frameMaterial: string;
  source: string;
  article: string;
  price: number;
  routerLink: string;
}